+++ internal/gw/helpers.go
@@ -4,10 +4,8 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 
-	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/hlog"
 )
 
@@ -46,42 +44,6 @@ func (app *application) getRawAccessToken(r *http.Request) (string, error) {
 	}
 
 	return "", fmt.Errorf("no bearer token found")
-}
-
-// lshortfile implements Lshortfile equivalent for zerolog's CallerMarshalFunc.
-func (app *application) lshortfile(file string, line int) string {
-	// Copied from the standard library: https://cs.opensource.google/go/go/+/refs/tags/go1.17.8:src/log/log.go;drc=926994fd7cf65b2703552686965fb05569699897;l=134
-	short := file
-	for i := len(file) - 1; i > 0; i-- {
-		if file[i] == '/' {
-			short = file[i+1:]
-			break
-		}
-	}
-	file = short
-	return file + ":" + strconv.Itoa(line)
-}
-
-// enrichLogContext adds a custom field and a value to zerolog context.
-func (app *application) enrichLogContext(r *http.Request, field string, value string) {
-	if field != "" && value != "" {
-		log := zerolog.Ctx(r.Context())
-		log.UpdateContext(func(c zerolog.Context) zerolog.Context {
-			return c.Str(field, value)
-		})
-	}
-}
-
-// enrichDebugLogContext adds a custom field and a value to zerolog context if logging level is set to Debug.
-func (app *application) enrichDebugLogContext(r *http.Request, field string, value string) {
-	if app.Debug {
-		if field != "" && value != "" {
-			log := zerolog.Ctx(r.Context())
-			log.UpdateContext(func(c zerolog.Context) zerolog.Context {
-				return c.Str(field, value)
-			})
-		}
-	}
 }
 
 // isNotAPIRequest returns true if the requested path does not target API or federate endpoints.
